Split report writing from test running. Add emoji and relative links
